% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_ili}
\alias{sample_ili}
\title{Simulated ILI Surveillance Data}
\format{
A data frame with 200 rows and 2 variables:
\itemize{
  \item date: Date of observation (weekly)
  \item case: Integer count of reported cases
}
}
\usage{
data(sample_ili)
}
\description{
A dataset containing 200 weeks of simulated influenza-like illness case counts.
}
